[ios] Add `Trips.csv` to `Resources`
We're including all 1.6MB for now, but we'll want to replace with some derived table that drops extraneous data. Starting with this because this is the shape of data we'll receive from endpoint when we fetch dynamically, so we'll need to "fetch" 1.6MB of data over network regardless of how we persist. Unless we serve the derived data ourselves... which doesn't make sense right now