[ios] Implement `getStationsWith(id: Int) -> [MTAStation]`
Super simple, need this to get other `MTAStation` with same `stationID` (BUT different `gtfsID`).

For example, `A32` is "W 4 St-Wash Sq" (stationID: 167) GTFS ID corresponding to ACE, but `D20` is "W 4 St-Wash Sq" (stationID: 167) GTFS ID corresponding to BDFM.

We want to coalesce `MTAStation` with same `stationID` together... which is def a little confusing. `MTAStation` should probably be renamed `MTAStop` or something to be more accurate.