[ios] Introduce `getArrivalsFor(station: MTAStation) async -> [TrainArrivalEntry] `
This reduces duplication in `ContentView`.

It's not great API design that we have a separate `getArrivalsFor(station: MTAStation)` and `getTrainArrivalsForStop(stop, feed)`.

It's super weird that we have a function that requires passing in GRPC feed, that should be so abstracted far far away. Will tidy up once more things are working since it's invisible to user.