[ios] Update `Trips.csv` with `shape_id` field patched in
Manually ran `PatchTripsCSV` script on Trips.csv that was manually created by concatenating the base and supplemental GTFS static data.

This will fix things for now, but is not a great approach going forward. We'll want to programmatically fetch both files, concatenate them, and patch in missing data... but we'll defer that for now.