[ios] Update `shapeToTerminus` with `tripID` keys instead of `shapeID`
This resolves issues with ambiguous terminal station labels. This is not a great solution, but it should be workable for now.

Want to get the data 100% correct and accurate before optimizing. We could do something where we "compress" dictionary by collapsing keys where common prefix points to unique terminal station or something like that, but probably easier to just dump all the Trips in SQLite and query on demand to avoid keeping all this junk in memory.