[ios] Reintroduce `shapeToTerminusFromCSV`
We're going to go with a hybrid of `shapeID`s, `tripID`s, and partial matches of both for the timebeing. That **should** cover all of the cases, but we'll monitor with `terminal_station_mismatch` events in PostHog.

If using those doesn't cover 100%, we should try to get the headsign from the `NYCTTripDescriptor` which was being difficult before. If we still can't get 100% of IDs to match, will probably make sense to look at some of the open source stuff on GitHub to see how they're handling all scenarios. Would also be good to have some sort of CI thing that runs periodically that compares our results (as function of static, supplemental static, and realtime) with that of "internal"(?) endpoint that's used on eg https://map.mta.info/.