[ios] Update `determineTerminalStation(...)` to handle more scenarios.
1. We see if there is an exact match in `tripIDToTerminus`.
2. We see if there is an exact match in `shapeIDToTerminus`.
3. We see if there is a partial match in `shapeIDToTerminus`.

Hopefully this brings the number of mismatches to 0. However, if we still aren't able to get terminal station with these checks we should be able to with `NyctTripDescriptor`. The issue is that I've been unable to pull `NyctTripDescriptor` out of the GRPC response for some weird reason...

But at least there is another obvious step if this doesn't fully resolve issue. It might be good to include the `NyctTripDescriptor` check for completeness eventually, but not a priority if mismatches stop showing up.