[ios] Rename `Flatbush Av-Brooklyn College` to `Flatbush-Brooklyn College`
This is the only terminal station that extends to multiple lines in `ArrivalCard` 😭. I tried so many ways of adjusting layout just to accomodate this on a single line, but nothing looked good.

I tried cutting the text off with ellipsis, putting in horizontal scroll view, animating horizontal scroll, messing with padding and layout, changing font sizes, etc... but everything was strictly worse for all `ArrivalCard`s.

So we're going to take a shortcut and just cut the " Av" so that it all fits on one line (on my iPhone 14 Pro that is...). This is so hacky and I don't love that I'm messing with the sanctity of the data for my own selfish/lazy reasons... but who cares.