[Scripts] Add `processTripsFile(inputURL: URL, outputURL: URL)` step to `FetchStaticData`
This drops `service_id` column which we're not using and drops instance-specific tripID prefix (eg `ASP24GEN-1038-Sunday-00_`) which we can cut.

We can now look at dropping rows with duplicate tripIDs and then also cleaning up files... though it's probably not a big deal if we run this via GitHub Action.