[ios] Add `.uniqued(on: \.id)` to `arrivalEntries`
So we don't have duplicate entries in the case of eg Fulton Street where there are 123 and 456 trains running.

In an ideal world we wouldn't be making multiple requests and cosntructing multiple arrival entries to begin with, but this solves id stability issue in ListView for now. Will look at the fetching code later.