[Sources] Populate SwiftData `TripEntry` model container in `ChooChoo.init()`
Right now we're filling the DB with everything in `Trips.csv` every single time the app loads. This is obviously not optimal.

We could store some boolean in `UserDefaults` like `hasInitialDataBeenLoaded` or something so we only do this on first launch and "fill in" newer data from network later on?

Let's just get this working for now and refine approach later.