[Sources] Greatly simplify logic to determine terminal station
Wow... I was under the incorrect assumption that we were NOT guaranteed to have `stopTimeUpdate`s that include entries for every stop in an active trip in the GTFS-RT feed. And this is TRUE for many transit agencies publishing GTFS-RT data. They only have the arrival/departure estimates for the next or next few stations and you need to use `trips.txt` and `routes.txt` and `stations.txt` and `calendar.txt` to stitch everything together and approximate.

HOWEVER, based on https://groups.google.com/g/onebusaway-developers/c/KD4Qijjbx8w/m/tjbRedMEV5EJ, the MTA guarantees that all stops on an active trip WILL be included in the `stopTimeUpdate`.....

So we can literally just do `stopTimeUpdate.last.gtfsID` and get the terminal station for an active trip.

There has been so much wasted code and effort trying to match on `tripID`s, partial `tripID`s, `shapeID`s, partial `shapeID`s, etc. that we'll be able to cut out. Which is annoying.

On the plus side, this greatly simplifies code, removes need for our handling lifecycle of static data in the app, etc etc... so can get back to interesting features instead of this stupid problem.