[Sources] Introduce skeleton `MTAStopValue`
We're creating a value type version of `MTAStop` to be passed to the `Task` in `ContentView`.

There really has to be a better way to offload work from main thread when working with SwiftData objects that doesn't require "detaching" like this, but this will work for now. I'll watch some WWDC videos later, there are other things to prioritize at the moment though.