[Sources] Tidy up `AlertIndicator` and place in `StationSign`
Spent a lot of time trying different things with the layout here.

It's challenging because the visual symmetry of every station sign is so different. Some have long names and 1-2 lines, some have short names and 6+ lines... and everything in between.

I think for now I'm going to go with "static" informational text on left side of auxillary bar, and "actions" on the right. Wonder if we should do a left-handed mode where it's flipped for button reachability?