[Sources] Introduce `AlertSheet` and surface when `AlertIndicator` is tapped
We definitely need to work on styling and layout here... but at least all the data is in the right place!

This is the minimally release-able version with alerts I think. Will still try to get a few more changes in, would be nice to submit a build tonight that's hopefully ready by tomorrow morning.