[Sources] Introduce `parseAlertBody(...)` to render `TrainBadge`s in service alert text
Surprised this work. Basically worked backwords from the manual concatenation approach that I got working with the "Delays on the [F] and [G] trains example. Thank god we don't need to use `HStack` or whatever and deal with all the text layout weirdness.

I'm not sure if this is the most performant approach, but it works for now so will leave it.